functions/resize/requirements.txt
pillow
minio